+++ app/src/main/java/sensors_in_paradise/sonar/screen_data/TrainingHistoryAdapter.kt
@@ -1,4 +1,4 @@
-package sensors_in_paradise.sonar.screen_train
+package sensors_in_paradise.sonar.screen_data
 
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
